refactored route folder structure
