+++ routes/access/access.js
@@ -1,0 +1,3 @@
+module.exports = (req, res) => {
+    res.json("Success");
+  };
