+++ routes/index.js
@@ -1,0 +1,8 @@
+const routes = require('express').Router();
+const login = require('./login');
+const access = require('./access');
+
+routes.use('/login', login);
+routes.use('/access', access);
+
+module.exports = routes;
