+++ routes/login/index.js
@@ -1,0 +1,6 @@
+const main = require('express').Router();
+const login = require('./login');
+
+main.post('/', login);
+
+module.exports = main;
